pyproject.toml
[project]
name = "job-application-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "pydantic[dotenv]>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sqlalchemy[asyncio]>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]
